requirements.txt
// ai_prompt_eval/
// ├── .env
// ├── requirements.txt
// ├── src/
// │   ├── __init__.py
// │   ├── main.py
// │   ├── models/
// │   │   ├── __init__.py
// │   │   └── gemini_model.py
// │   └── prompts/
// │       ├── system_prompt.txt
// │       └── user_prompt.txt
langchain
langchain-google-genai
google-generativeai
python-dotenv
langchain-community
langchain-openai # 添加这一行
# requests # 如果其他地方不用可以移除
lark-oapi # 添加这一行
